TPG/requirements.txt
Flask==3.0.3
pyserial==3.5
pandas==2.2.2
openpyxl==3.1.2
pyngrok==7.1.6
requests==2.31.0
python-dotenv==1.0.1
